move release preview pill
