+++ products/error_tracking/frontend/components/ReleasesPreview/ReleasePreviewPill.tsx
@@ -1,4 +1,3 @@
-import { useValues } from 'kea'
 import { useState } from 'react'
 
 import { IconCommit } from '@posthog/icons'
@@ -7,16 +6,9 @@ import { LemonTag, Popover } from '@posthog/lemon-ui'
 import { ErrorTrackingRelease } from 'lib/components/Errors/types'
 
 import { ReleasePopoverContent } from './ReleasesPopoverContent'
-import { releasePreviewLogic } from './releasePreviewLogic'
 
-export function ReleasePreviewPill(): JSX.Element {
+export function ReleasePreviewPill({ release }: { release: ErrorTrackingRelease }): JSX.Element {
     const [isOpen, setIsOpen] = useState(false)
-    const { release } = useValues(releasePreviewLogic)
-
-    if (!release) {
-        return <></>
-    }
-
     return (
         <Popover
             visible={isOpen}
