+++ products/error_tracking/frontend/components/ReleasesPreview/ReleasesPopoverContent.tsx
