+++ products/error_tracking/frontend/components/ReleasesPreview/gitMetadataParser.spec.ts
