+++ products/error_tracking/frontend/components/ReleasesPreview/gitMetadataParser.ts
